refactor(circular-deps): break ducks/metamask ↔ store/actions cycle
Extract `updateTransactionGasFees` into its own module under
`ui/store/actions/update-transaction-gas-fees.ts` and re-export it from
`ui/store/actions.ts` for back-compat. The metamask duck now imports
the action directly from the new file, breaking the cycle.

No runtime change — pure file/import reorganization.
