+++ .madgerc
@@ -25,7 +25,6 @@
   "allowedCircularGlob": [
     "ui/ducks/**",
     "ui/selectors/**",
-    "ui/store/**",
     "ui/components/app/**"
   ]
 }
